Image saved Successfully

+++ Frontend_Angular_Crud_Operation/src/app/student/student.component.ts
@@ -1,6 +1,9 @@
 import { StudentService } from './../student.service';
 import { Component } from '@angular/core';
 import { Student } from '../student';
+import Swal from 'sweetalert2';
+
+
 
 @Component({
   selector: 'app-student',
@@ -13,14 +16,38 @@ export class StudentComponent {
   studentlist:Student[]=[];
   newStudent:Student=new Student();
   editStudent:Student=new Student();
-  
+  imageStudent:Student=new Student();
+  img:any;
   constructor(private studservice:StudentService){}
-  
   ngOnInit()
   {
     this.Getall();
   }
- 
+  onFileSelected(event: any) {
+   
+    const file: File = event.target.files[0];
+    const reader = new FileReader();
+    reader.readAsDataURL(file);
+    reader.onload = (event:any) => {
+      this.img=event.target.result;
+      const base64Image: any = reader.result as string
+      console.log(base64Image); // This is the base64 encoded image
+      this.imageStudent.image=base64Image
+      this.editStudent.image=base64Image
+    };
+  }
+
+//   saveImage()
+//   {
+//     this.studservice.saveimage(this.imageStudent)
+     
+// this.Getall
+// this.newStudent.name="";
+// this.newStudent.address="";
+// this.newStudent.phoneNumber="";
+      
+//   }
+
   Getall()
   {
     this.studservice.getStudent().subscribe(
@@ -33,14 +60,15 @@ export class StudentComponent {
       }
     );
   }
-  Save(){
+  saveimage(){
     
-    this.studservice.postStudent(this.newStudent).subscribe(
+    this.studservice.postStudent(this.imageStudent).subscribe(
       (response)=>{
 this.Getall();
-this.newStudent.name="";
-this.newStudent.address="";
-this.newStudent.phoneNumber="";
+this.imageStudent.name="";
+this.imageStudent.address="";
+this.imageStudent.phoneNumber="";
+this.imageStudent.image="";
       }
     )
   }
@@ -61,17 +89,41 @@ this.newStudent.phoneNumber="";
       }
     )
   }
+
   deleteclick(id:number)
   {
-    alert(id);
-    this.studservice.DeleteStudent(id).subscribe(
-      (response)=>{
-      this.Getall();
-    },
-    (error)=>{
-      console.log(error)
-    }
-    );
+    Swal.fire({  
+      title: 'Are you sure want to remove?',  
+      text: 'You will not be able to recover this file!',  
+      icon: 'warning',  
+      showCancelButton: true,  
+      confirmButtonText: 'Yes, delete it!',  
+      cancelButtonText: 'No, keep it'  
+    }).then((result) => {  
+      if (result.value) {  
+        this.studservice.DeleteStudent(id).subscribe(
+      
+          (response)=>{
+            
+         
+          this.Getall();
+        }
+        )
+        Swal.fire(  
+          'Deleted!',  
+          'Your imaginary file has been deleted.',  
+          'success'  
+        )  
+      } else if (result.dismiss === Swal.DismissReason.cancel) {  
+        Swal.fire(  
+          'Cancelled',  
+          'Your imaginary file is safe :)',  
+          'error'  
+        )  
+      }  
+    })  
+   
   }
+
 }
